Refactor distribution analysis to support floating-point data and modularize the result UI.

Update EngineService, LocalEngineServiceImpl, and CalculationSystemRepository to accept DoubleArray instead of IntArray for distribution calculations.
Update DistributionParams and MainScreenEvent to use List<Double> for input data.
Modify DistributionForm logic to parse input sequences as doubles rather than integers.
Extract distribution result layout from ResultForm.kt into a standalone DistributionResultScreen component.
Enhance SingleResultCard and ComplexResultCard with improved styling, including inner shadows, rounded corners, and consistent background colors.
Add fillMaxSize modifier to CustomEditText for better layout consistency.

+++ app/src/main/java/com/example/stochia/repository/CalculationSystemRepository.kt
@@ -5,7 +5,7 @@ import com.example.stochia.core.services.calculation_system.local.LocalEngineSer
 
 class CalculationSystemRepository(private val engineService: LocalEngineServiceImpl) {
 
-    fun getDistribution(data: IntArray): PyObject {
+    fun getDistribution(data: DoubleArray): PyObject {
         return engineService.getDistribution(data)
     }
